removed backend folder and replaced with core, added PDF, PPT, and JPG processing methods, added test files to test methods, added python dependencies to requirements.txt
