+++ packages/interface/requirements.txt
@@ -1,2 +1,5 @@
 fastapi==0.115.4
 sentence-transformers==3.2.1
+PyPDF2==3.0.1
+python-pptx==1.0.2
+pytesseract==0.3.13
